# internal/store/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"time"

	"github.com/google/uuid"
)

type Answer struct {
	ID          uuid.UUID `json:"id"`
	AttemptID   uuid.UUID `json:"attempt_id"`
	TaskID      uuid.UUID `json:"task_id"`
	RawAnswer   string    `json:"raw_answer"`
	IsCorrect   bool      `json:"is_correct"`
	TimeSpentMs int64     `json:"time_spent_ms"`
	AnsweredAt  time.Time `json:"answered_at"`
}

type Assignment struct {
	ID          uuid.UUID  `json:"id"`
	TestID      uuid.UUID  `json:"test_id"`
	StudentID   uuid.UUID  `json:"student_id"`
	AssignedBy  uuid.UUID  `json:"assigned_by"`
	ScheduledAt time.Time  `json:"scheduled_at"`
	NotifiedAt  *time.Time `json:"notified_at"`
	Status      string     `json:"status"`
}

type Attempt struct {
	ID           uuid.UUID  `json:"id"`
	AssignmentID *uuid.UUID `json:"assignment_id"`
	TestID       uuid.UUID  `json:"test_id"`
	StudentID    uuid.UUID  `json:"student_id"`
	StartedAt    time.Time  `json:"started_at"`
	FinishedAt   *time.Time `json:"finished_at"`
}

type Enrollment struct {
	ID        uuid.UUID `json:"id"`
	TeacherID uuid.UUID `json:"teacher_id"`
	StudentID uuid.UUID `json:"student_id"`
}

type Subject struct {
	ID    uuid.UUID `json:"id"`
	Code  string    `json:"code"`
	Title string    `json:"title"`
}

type Task struct {
	ID           uuid.UUID `json:"id"`
	SubjectID    uuid.UUID `json:"subject_id"`
	Number       int32     `json:"number"`
	Statement    string    `json:"statement"`
	Media        []byte    `json:"media"`
	AnswerSchema []byte    `json:"answer_schema"`
	Source       []byte    `json:"source"`
	Status       string    `json:"status"`
	CreatedAt    time.Time `json:"created_at"`
}

type Test struct {
	ID        uuid.UUID `json:"id"`
	SubjectID uuid.UUID `json:"subject_id"`
	Kind      string    `json:"kind"`
	Title     string    `json:"title"`
	CreatedBy uuid.UUID `json:"created_by"`
	CreatedAt time.Time `json:"created_at"`
}

type TestItem struct {
	ID       uuid.UUID `json:"id"`
	TestID   uuid.UUID `json:"test_id"`
	TaskID   uuid.UUID `json:"task_id"`
	Position int32     `json:"position"`
}

type User struct {
	ID           uuid.UUID `json:"id"`
	Role         string    `json:"role"`
	TelegramID   *int64    `json:"telegram_id"`
	Name         string    `json:"name"`
	CreatedAt    time.Time `json:"created_at"`
	Username     *string   `json:"username"`
	PasswordHash *string   `json:"password_hash"`
}